Rename endpoint /epci to /epcis to align with other endpoints convention
